Brawler v2.0

+++ Assets/_Scripts/CreatureBehaviour/Enemy/IDamageable.cs
@@ -1,8 +1,8 @@
 public interface IDamageable
 {
-    int MaxHealth { get; set; }
-    int CurrentHealth { get; set; }
+    float MaxHealth { get; set; }
+    float CurrentHealth { get; set; }
     
-    void TakeDamage(int damage);
+    void TakeDamage(float damage);
     void Die();
 }
